rename snapshot -> snapshot_indexer
