+++ src/lib/codegen/canisters/snapshot_json_rpc.rs
@@ -2,7 +2,7 @@ use quote::{format_ident, quote};
 
 use crate::lib::codegen::components::snapshot_json_rpc::SnapshotJsonRPCComponentManifest;
 
-use super::snapshot::generate_queries_without_timestamp;
+use super::snapshot_indexer::generate_queries_without_timestamp;
 
 pub fn generate_codes(
     manifest: &SnapshotJsonRPCComponentManifest,
