+++ src/lib/codegen/components/snapshot_indexer.rs
@@ -47,7 +47,7 @@ impl SnapshotComponentManifest {
             version: version.to_owned(),
             metadata: ComponentMetadata {
                 label: label.to_owned(),
-                type_: ComponentType::Snapshot,
+                type_: ComponentType::SnapshotIndexer,
                 description: description.to_owned(),
                 tags: Some(vec![
                     "ERC-20".to_string(),
@@ -77,22 +77,22 @@ impl ComponentManifest for SnapshotComponentManifest {
     }
 
     fn validate_manifest(&self) -> anyhow::Result<()> {
-        canisters::snapshot::validate_manifest(self)
+        canisters::snapshot_indexer::validate_manifest(self)
     }
 
     fn generate_codes(
         &self,
         _interface_contract: Option<ethabi::Contract>,
     ) -> anyhow::Result<TokenStream> {
-        canisters::snapshot::generate_codes(self)
+        canisters::snapshot_indexer::generate_codes(self)
     }
 
     fn generate_scripts(&self, network: Network) -> anyhow::Result<String> {
-        scripts::snapshot::generate_scripts(self, network)
+        scripts::snapshot_indexer::generate_scripts(self, network)
     }
 
     fn component_type(&self) -> ComponentType {
-        ComponentType::Snapshot
+        ComponentType::SnapshotIndexer
     }
 
     fn metadata(&self) -> &ComponentMetadata {
@@ -209,7 +209,7 @@ interval: 3600
                 version: "v1".to_owned(),
                 metadata: ComponentMetadata {
                     label: "sample_pj_snapshot_chain".to_owned(),
-                    type_: ComponentType::Snapshot,
+                    type_: ComponentType::SnapshotIndexer,
                     description: "Description".to_string(),
                     tags: Some(vec!["ERC-20".to_string(), "Ethereum".to_string()])
                 },
@@ -270,7 +270,7 @@ interval: 3600
                 version: "v1".to_owned(),
                 metadata: ComponentMetadata {
                     label: "sample_pj_snapshot_icp".to_owned(),
-                    type_: ComponentType::Snapshot,
+                    type_: ComponentType::SnapshotIndexer,
                     description: "Description".to_string(),
                     tags: Some(vec!["ERC-20".to_string(), "Ethereum".to_string()])
                 },
