+++ src/lib/codegen/scripts/mod.rs
@@ -3,5 +3,5 @@ pub mod algorithm_lens;
 pub mod common;
 pub mod event_indexer;
 pub mod relayer;
-pub mod snapshot;
+pub mod snapshot_indexer;
 pub mod snapshot_json_rpc;
